requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_sysroot_linux-64_curr_repodata_hack=3=haa98f57_10
absl-py=2.1.0=py310h06a4308_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.15=pypi_0
aiosignal=1.4.0=pypi_0
aom=3.6.0=h6a678d5_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
binutils_impl_linux-64=2.40=h5293946_0
binutils_linux-64=2.40.0=hc2dff05_2
blas=1.0=mkl
blessed=1.21.0=pypi_0
bottleneck=1.4.2=py310ha9d4c09_0
brotli-python=1.0.9=py310h6a678d5_9
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
cairo=1.16.0=he5ede1b_6
certifi=2025.8.3=pypi_0
charset-normalizer=3.4.2=pypi_0
cloudpickle=3.1.1=py310h06a4308_0
contourpy=1.3.1=py310hdb19cb5_0
cuda-cccl_linux-64=12.4.127=h06a4308_2
cuda-crt-dev_linux-64=12.4.131=h06a4308_0
cuda-crt-tools=12.4.131=h06a4308_0
cuda-cudart=12.4.127=h99ab3db_0
cuda-cudart-dev=12.4.127=h99ab3db_0
cuda-cudart-dev_linux-64=12.4.127=hd681fbe_0
cuda-cudart-static=12.4.127=h99ab3db_0
cuda-cudart-static_linux-64=12.4.127=hd681fbe_0
cuda-cudart_linux-64=12.4.127=hd681fbe_0
cuda-cuobjdump=12.4.127=h6a678d5_1
cuda-cupti=12.4.127=h6a678d5_1
cuda-driver-dev_linux-64=12.4.127=hd681fbe_0
cuda-nvcc=12.4.131=h02f8991_0
cuda-nvcc-dev_linux-64=12.4.131=h4ee8466_0
cuda-nvcc-impl=12.4.131=h99ab3db_0
cuda-nvcc-tools=12.4.131=h99ab3db_0
cuda-nvcc_linux-64=12.4.131=he92618c_0
cuda-nvdisasm=12.4.127=h6a678d5_1
cuda-nvrtc=12.4.127=h99ab3db_1
cuda-nvtx=12.4.127=h6a678d5_1
cuda-nvvm-dev_linux-64=12.4.131=h06a4308_0
cuda-nvvm-impl=12.4.131=h6a678d5_0
cuda-nvvm-tools=12.4.131=h6a678d5_0
cuda-version=12.4=hbda6634_3
cudnn=9.1.1.17=cuda12_1
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h1110e0f_3
dav1d=1.2.1=h5eee18b_0
deprecated=1.2.13=py310h06a4308_0
expat=2.7.1=h6a678d5_0
farama-notifications=0.0.4=py310h06a4308_0
ffmpeg=6.1.1=h2a67f75_3
filelock=3.17.0=py310h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.55.3=py310h5eee18b_0
freetype=2.13.3=h4a9f257_0
frozenlist=1.7.0=pypi_0
fsspec=2025.3.2=py310he3bba80_0
gcc_impl_linux-64=11.2.0=h1234567_1
gcc_linux-64=11.2.0=h5c386dc_2
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py310h5eee18b_0
gpustat=1.1.1=pypi_0
graphite2=1.3.14=h295c915_1
grpcio=1.71.0=py310h6a678d5_0
gxx_impl_linux-64=11.2.0=h1234567_1
gxx_linux-64=11.2.0=hc2dff05_2
gymnasium=1.1.1=pypi_0
harfbuzz=10.2.0=hdfddeaa_1
icu=73.1=h6a678d5_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jax-jumpy=1.0.0=pyhd8ed1ab_1
jinja2=3.1.6=py310h06a4308_0
jpeg=9e=h5eee18b_3
kernel-headers_linux-64=3.10.0=h57e8cba_10
kiwisolver=1.4.8=py310h6a678d5_0
lame=3.100=h7b6447c_0
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.40=h12ee557_0
leptonica=1.82.0=hfdeec58_3
lerc=4.0.0=h6a678d5_0
libabseil=20250127.0=cxx17_h6a678d5_0
libarchive=3.7.7=hfab0078_0
libcublas=12.4.5.8=h99ab3db_1
libcufft=11.2.1.3=h99ab3db_1
libcurand=10.3.5.147=h99ab3db_1
libcusolver=11.6.1.9=h99ab3db_1
libcusparse=12.3.1.170=h99ab3db_1
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-devel_linux-64=11.2.0=h1234567_1
libgcc-ng=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libgrpc=1.71.0=h2d74bed_0
libiconv=1.16=h5eee18b_3
libkrb5=1.21.3=h520c7b4_3
libllvm19=19.1.7=he89c38a_1
libmlir19=19.1.7=hdb19cb5_0
libnvjitlink=12.4.127=h99ab3db_1
libnvjpeg=12.3.1.117=h6a678d5_1
libogg=1.3.5=h27cfd23_1
libopus=1.3.1=h5eee18b_1
libpng=1.6.39=h5eee18b_0
libpq=17.4=h02b6914_2
libprotobuf=5.29.3=hc99497a_0
libre2-11=2024.07.02=h6a678d5_0
libstdcxx-devel_linux-64=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libtheora=1.1.1=h7f8727e_3
libtiff=4.7.0=hde9077f_0
libtorch=2.5.1=gpu_cuda124_hf726350_304
libuuid=1.41.5=h5eee18b_0
libuv=1.48.0=h5eee18b_0
libvorbis=1.3.7=h7b6447c_0
libvpx=1.13.1=h6a678d5_0
libwebp=1.3.2=h9f374a3_1
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.17.0=h9b100fa_0
libxml2=2.13.8=hfdd30dd_0
lit=19.1.7=py310h06a4308_1
lmdb=0.9.29=h2531618_0
lz4-c=1.9.4=h6a678d5_1
magma=2.7.1=ha545f22_0
markdown=3.8=py310h06a4308_0
markupsafe=3.0.2=py310h5eee18b_0
matplotlib-base=3.10.0=py310hbfdbfaf_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_2
mkl_fft=1.3.11=py310h5eee18b_0
mkl_random=1.2.8=py310h1128e8f_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py310h06a4308_0
multidict=6.6.3=pypi_0
nccl=2.21.5.1=ha515578_0
ncurses=6.4=h6a678d5_0
networkx=3.4.2=py310h06a4308_0
numexpr=2.11.0=py310hf4471f9_0
numpy=2.0.1=py310h5f9d8c6_1
numpy-base=2.0.1=py310hb5e798b_1
nvidia-ml-py=13.580.65=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=h0d4d230_1
openldap=2.6.4=h42fbc30_0
openssl=3.0.17=h5eee18b_0
opentelemetry-api=1.30.0=py310h06a4308_0
packaging=25.0=py310h06a4308_0
pandas=2.3.1=py310h277b779_0
pcre2=10.42=hebb0a14_1
pillow=11.1.0=py310hac6e08b_1
pip=25.1=pyhc872135_2
pixman=0.40.0=h7f8727e_1
propcache=0.3.2=pypi_0
protobuf=5.29.3=py310he621ea3_0
psutil=7.0.0=pypi_0
psycopg2=2.9.10=py310h5eee18b_0
pthread-stubs=0.3=h0ce48e5_1
pyparsing=3.2.0=py310h06a4308_0
python=3.10.18=h1a3bd86_0
python-dateutil=2.9.0post0=py310h06a4308_2
python-tzdata=2025.2=pyhd3eb1b0_0
pytorch=2.5.1=gpu_cuda124_py310hfee4293_304
pytz=2025.2=py310h06a4308_0
re2=2024.07.02=hdb19cb5_0
readline=8.2=h5eee18b_0
requests=2.32.4=pypi_0
setuptools=78.1.1=py310h06a4308_0
six=1.17.0=py310h06a4308_0
sleef=3.5.1=h5eee18b_2
sqlglot=27.13.2=pypi_0
sqlglotrs=0.6.2=pypi_0
sqlite=3.45.3=h5eee18b_0
sqlparse=0.5.3=pypi_0
stable-baselines3=2.6.0=pypi_0
sympy=1.13.3=py310h06a4308_1
sysroot_linux-64=2.17=h57e8cba_10
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.20.0=pyhe01879c_0
tensorboard-data-server=0.7.0=py310h52d8a92_1
tesseract=5.2.0=h6a678d5_2
tk=8.6.14=h993c535_1
torch-geometric=2.6.1=pypi_0
torchinfo=1.8.0=pypi_0
torchvision=0.20.1=cuda124py310ha0e159c_102
tqdm=4.67.1=py310h2f386ee_0
triton=3.1.0=cuda124py310hac27663_3
typing-extensions=4.12.2=py310h06a4308_0
typing_extensions=4.12.2=py310h06a4308_0
tzdata=2025b=h04d1e81_0
unicodedata2=15.1.0=py310h5eee18b_1
urllib3=2.5.0=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.1.3=py310h06a4308_0
wheel=0.45.1=py310h06a4308_0
wrapt=1.17.0=py310h5eee18b_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-libxext=1.3.6=h9b100fa_0
xorg-libxrender=0.9.12=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yarl=1.20.1=pypi_0
zipp=3.21.0=py310h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
